requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
python-docx>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0